refactor: use pkg
